requirements_chatbot.txt
langchain==0.3.24
python-dotenv==1.0.0
streamlit==1.22.0
tiktoken==0.9.0
faiss-cpu==1.10.0
protobuf~=3.19.0
langchain-google-genai~=2.1.3
langchain-community~=0.3.22
sentence-transformers~=4.1.0